made some code refactors
