+++ apps/web/app/s/[surveyId]/LegalFooter.tsx
