+++ apps/web/app/s/[surveyId]/page.tsx
@@ -2,14 +2,14 @@ export const revalidate = REVALIDATION_INTERVAL;
 
 import { REVALIDATION_INTERVAL } from "@formbricks/lib/constants";
 import LinkSurvey from "@/app/s/[surveyId]/LinkSurvey";
-import LegalFooter from "@/components/shared/LegalFooter";
+import LegalFooter from "@/app/s/[surveyId]/LegalFooter";
 import { getSurvey } from "@formbricks/lib/services/survey";
 import { getProductByEnvironmentId } from "@formbricks/lib/services/product";
 import SurveyInactive from "@/app/s/[surveyId]/SurveyInactive";
 
 export default async function LinkSurveyPage({ params }) {
-  const survey = await getSurvey(params.surveyId);
-  const product = await getProductByEnvironmentId(params.environmentId);
+
+  const [survey, product] = await Promise.all([getSurvey(params.surveyId), getProductByEnvironmentId(params.environmentId)]);
 
   if (survey && survey.status !== "inProgress") {
     return <SurveyInactive status={survey.status} surveyClosedMessage={survey.surveyClosedMessage} />;
